producer/requirements.txt
kafka-python
request
psycopg2-binary
python-dotenv